fashion_store/requirements.txt
asgiref==3.6.0
certifi==2020.6.20
chardet==3.0.4
charset-normalizer==3.1.0
Django==3.2
idna==3.4
mysqlclient==2.1.1
passlib==1.7.2
Pillow==9.5.0
pytz==2020.1
requests==2.28.2
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
